Various cleanups
Remove the `DomainID` field from the `Provider` struct and instead
just look for the Linode domain whose `Domain` field matches the
passed-in `zone` argument in `Provider.getDomainIDByZone`.  Also, make
`Provider.getDomainIDByZone` a bit more readable.

Add a private `sync.Mutex` field to `Provider` and use it to
synchronize all exported method calls.

`Provider.SetRecords` now uses an empty `ID` field of a passed-in
`libdns.Record` to indicate that a create record should be made,
otherwise an existing record with the same ID is updated.  This avoids
error-prone comparisons with `Name` fields which we were doing
previously.

Cast record type to `linodego.DomainRecordType` since it's a `string`
anyways when converting Linode records to a `libdns.Record`.

Use `libdns.RelativeName` to ensure the `Name` field of a
`libdns.Record`'s construct contain a relative name.

Ensure we trim any trailing dot (`.`) at the end of zone names before
comparing them with Linode names which do not end with a dot.